+++ tests/testthat/test-drks_de_version.R
@@ -1,7 +1,7 @@
 test_that("DRKS.de version downloads correctly", {
     result <- drks_de_version("DRKS00005219", 1)
     expect_equal(
-        length(result) == 13 || result == "Error",
+        length(result) == 14 || result == "Error",
         TRUE
     )
 })